% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.interactions.R
\name{get.interactions}
\alias{get.interactions}
\title{Create network from social media data. Code is optimized for Twitter data but can be used with other data sources.
This function Transforms a tweet dataset into a network via the function get.interactions(). The function requires the object tweets (character vector of tweets) and the object screenName (character vector containing user names), along with a range of optional arguments to generate node (vertex) and edge (links) information, chief of which are timestamps and geographic information for nodes such as latitude and longitude coordinates. Along with the required arguments, the function accepts the following optional arguments and parameters:}
\usage{
get.interactions(
  tweets,
  screenName,
  node.attributes.df = NULL,
  edge.attributes.df = NULL,
  ascii = TRUE,
  return.network = "all.interactions",
  return.graph = TRUE,
  directed = TRUE,
  ...
)
}
\arguments{
\item{tweets:}{Character vector with tweets}

\item{screenName:}{Character vector with usernames}

\item{node.attributes.df:}{Data frame with node attributes. Defaults to NULL}

\item{edge.attributes.df:}{Data frame with edge attributes. Defaults to NULL}

\item{ascii:}{Whether text encoding is to be converted to ASCII. Defaults to TRUE}

\item{return.network:}{Network of "retweets," "mentions," or "all.interactions" (default)}

\item{return.graph:}{Return igraph object if TRUE (default) or data frame if FALSE}

\item{directed:}{Whether network is directed. Defaults to TRUE}
}
\value{
igraph object or edge list (data frame)
}
\description{
Create network from social media data. Code is optimized for Twitter data but can be used with other data sources. This function Transforms a tweet dataset into a network via the function get.interactions(). The function requires the object tweets (character vector of tweets) and the object screenName (character vector containing user names), along with a range of optional arguments to generate node (vertex) and edge (links) information, chief of which are timestamps and geographic information for nodes such as latitude and longitude coordinates. Along with the required arguments, the function accepts optional arguments and parameters.
}
