% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.spatial.network.R
\name{plot.spatial.network}
\alias{plot.spatial.network}
\title{The function plot.spatial.network() allows one to plot social network data over geographic grids using R base map library and the GADM database, a free spatial database of the location of the world’s administrative areas or boundaries. Network plots are generated using the capabilities of igraph, network, geosphere, ggplot2 (curved edges), or ggraph. The function accepts optional arguments and parameters.}
\usage{
plot.spatial.network(
  geo.graph,
  area = "UK",
  database = "world",
  package = "igraph",
  latitude = igraph::V(geo.graph)$place_lat,
  longitude = igraph::V(geo.graph)$place_lon,
  line.col = "black",
  nodeLabelnum = 5,
  vertex.size.igraph = as.numeric(log(igraph::degree(geo.graph) + 1) * 2),
  edge.width.igraph = "auto",
  edge.arrow.size.igraph = (50/igraph::ecount(geo.graph)) + 0.02,
  edge.color.igraph = igraph::E(geo.graph)$color,
  vertex.label.cex.igraph = 0.4,
  vertex.label.color.igraph = "black",
  vertex.cex.network = as.numeric(log(sna::degree(geo.n) + 1)/2),
  vertex.col.network = "orange",
  vertex.border.network = "black",
  usearrows.network = TRUE,
  arrowhead.cex.network = 0.2,
  label.cex.network = 0.5,
  edge.curve.network = 0.5,
  edge.width.ggplot = 10,
  alpha.ggplot = 1,
  edge.width.ggraph = "weight",
  avoid.overlap.ggraph = TRUE,
  vertex.label.cex.ggraph = 3,
  gadm.level = 2,
  ...
)
}
\arguments{
\item{geo.graph:}{network object}

\item{area:}{Geographic area to load base map. Defaults to UK}

\item{database:}{Database to load base map ("world" or "gadm"). Defaults to "world"}

\item{package:}{Preferred plotting package. Defaults to igraph}

\item{latitude:}{Latitude of nodes. Defaults to place_lat}

\item{longitude:}{Longitude of nodes. Defaults to place_lon}

\item{line.col:}{Color of line in the network plot. Defaults to black}

\item{nodeLabelnum:}{Node label size. Defaults to 5}

\item{vertex.size.igraph:}{Node size (igraph only). Defaults to a function of degree}

\item{edge.width.igraph:}{Width of edge (igraph only). Defaults to "auto"}

\item{edge.arrow.size.igraph:}{Arrow size (igraph only). Default is a function of density}

\item{edge.color.igraph:}{Edge color (igraph only). Defaults to edge attribute "color"}

\item{vertex.label.cex.igraph:}{Node label size (igraph only). Defaults to 0.4}

\item{vertex.label.color.igraph:}{Node label color (igraph only). Defaults to black}

\item{vertex.cex.network:}{Node size (network only). Defaults to a function of degree}

\item{vertex.col.network:}{Node color (network only). Defaults to orange}

\item{vertex.border.network:}{Node outline color (network only). Defaults to black}

\item{usearrows.network:}{(Logical) Use arrows in the plot. Defaults to TRUE}

\item{arrowhead.cex.network:}{Edge arrow size (network only). Defaults to 0.2}

\item{label.cex.network:}{Node label size (network only). Defaults to 0.5}

\item{edge.curve.network:}{Width of curved edge (network only). Defaults to 0.5}

\item{edge.width.ggplot:}{Width of edge (ggplot & geosphere). Defaults to 1}

\item{alpha.ggplot:}{Edge color alpha parameter (ggplot, geosphere, and ggraph). Defaults to 1}

\item{edge.width.ggraph:}{Width of edge (ggraph only). Defaults to node weight}

\item{avoid.overlap.ggraph:}{Avoid node label overlap (ggraph only). Defaults to TRUE}

\item{vertex.label.cex.ggraph:}{Node label size (ggraph only). Defaults to 3}

\item{gadm.level:}{Level of GADM base map to be loaded. Defaults to 2}
}
\value{
network plot over a geographic grid
}
\description{
The function plot.spatial.network() allows one to plot social network data over geographic grids using R base map library and the GADM database, a free spatial database of the location of the world’s administrative areas or boundaries. Network plots are generated using the capabilities of igraph, network, ggplot, ggplot2 (curved edges), or ggraph. The function accepts the following optional arguments and parameters:
}
